requirements.txt
gradio
langchain
langchain-community
faiss-cpu
sentence-transformers
pypdf
groq
ollama
python-dotenv
scikit-learn
tf-keras